+++ doc/CHANGES.rst
@@ -31,6 +31,7 @@ Futur
   with instances of the Models
 * FIX: create new session make must commit and remove all old session instances
 * IMP: add ``Mapping.clean`` method to clean unlinked mapping
+* IMP: add ``Mapping.remove_for_blokname`` method to remove mapping and obj
 
 0.9.7 (2017-07-03)
 ------------------